MDE/feature/revel Fixed revel pipeline
