+++ app/back_end/src/tools/__init__.py
@@ -17,7 +17,7 @@ from .spliceai import (
     add_spliceai_eval_columns,
 )
 
-from .revel.revel import (
+from .revel import (
     main_revel_pipeline
 )
 
@@ -35,5 +35,6 @@ __all__ = [
     "run_spliceai",
     "parse_spliceai_vcf",
     "add_spliceai_eval_columns",
+    # Revel related exports
     "main_revel_pipeline",
 ]
